+++ src/controllers/shop.controller.js
@@ -1,4 +1,5 @@
 const Shop = require("../models/shop.model");
+const User = require("../models/user.model");
 
 const shopController = {
   getAll: async (req, res) => {
@@ -23,8 +24,13 @@ const shopController = {
           },
         ],
       });
-      console.log(shop);
-      res.status(200).send(shop);
+      const user = req.user.id;
+      const isOwner = shop.user.toString() === user;
+      if (!isOwner) {
+        res.status(401).send({ message: "You are not the owner" });
+      } else {
+        res.status(200).send(shop);
+      }
     } catch (error) {
       res.status(400).send(error);
     }
@@ -32,7 +38,13 @@ const shopController = {
 
   create: async (req, res) => {
     try {
-      const shop = await Shop.create(req.body);
+      const user = req.user.id;
+      const userInDb = await User.findById(user);
+      if (userInDb.shop) {
+        return res.status(401).send({ message: "You already have a shop" });
+      }
+      const shop = await Shop.create({ ...req.body, user });
+      await User.findByIdAndUpdate(user, { shop: shop._id });
       res.status(201).send(shop);
     } catch (error) {
       res.status(400).send(error);
@@ -41,6 +53,11 @@ const shopController = {
 
   update: async (req, res) => {
     try {
+      const user = req.user.id;
+      const isOwner = shop.user.toString() === user;
+      if (!isOwner) {
+        res.status(401).send({ message: "You are not the owner" });
+      }
       const shop = await Shop.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
       });
@@ -52,7 +69,13 @@ const shopController = {
 
   delete: async (req, res) => {
     try {
+      const user = req.user.id;
+      const isOwner = shop.user.toString() === user;
+      if (!isOwner) {
+        return res.status(401).send({ message: "You are not the owner" });
+      }
       await Shop.findByIdAndDelete(req.params.id);
+
       res.status(200).send({ message: "Shop deleted" });
     } catch (error) {
       res.status(400).send(error);
